chore(wmg): refactor organism filter
